Remove Communications V3 API and stub v1 & v2 for the same
* Communications is an unused endpoint, so it’s not worth keeping this
debt until we actually start syncing Communications
