+++ .env.test
@@ -64,4 +64,4 @@ APPOINTMENT_NOTIFICATION_FACILITY_IDS=
 APPOINTMENT_NOTIFICATION_BOT_USER_UUID='a4fcb6b1-afb2-4ec1-952b-747754683497'
 APPOINTMENT_NOTIFICATION_BOT_USER_NAME='SMSReminder BotUser'
 APPOINTMENT_NOTIFICATION_HOUR_OF_DAY_START=0
-APPOINTMENT_NOTIFICATION_HOUR_OF_DAY_END=24
+APPOINTMENT_NOTIFICATION_HOUR_OF_DAY_FINISH=24
