+++ app/controllers/api/v1/communications_controller.rb
@@ -1,4 +1,13 @@
 class Api::V1::CommunicationsController < Api::V2::CommunicationsController
   include Api::V1::ApiControllerOverrides
-  include Api::V1::SyncControllerOverrides
+
+  def sync_to_user
+    render(
+      json: {
+        'communications' => [],
+        'processed_since' => Time.new(0)
+      },
+      status: :ok
+    )
+  end
 end
