--- app/controllers/api/v2/sync_controller.rb
@@ -1,2 +1,0 @@
-class Api::V2::SyncController < Api::Current::SyncController
-end
