+++ app/schema/api/v2/models.rb
@@ -21,5 +21,13 @@ class Api::V2::Models < Api::Current::Models
         :patient_business_identifiers,
       )
     end
+
+    def communication
+      super.tap do |d|
+        d[:properties][:communication_result][:enum] -= %i(unsuccessful unknown in_progress)
+      end.tap do |d|
+        d[:properties][:communication_type][:enum] -= %w(missed_visit_sms_reminder)
+      end
+    end
   end
 end
