+++ app/validators/api/v2/communication_payload_validator.rb
@@ -1,2 +1,16 @@
 class Api::V2::CommunicationPayloadValidator < Api::Current::CommunicationPayloadValidator
+  attr_accessor(
+    :id,
+    :appointment_id,
+    :user_id,
+    :communication_type,
+    :created_at,
+    :updated_at
+  )
+
+  validate :validate_schema
+
+  def schema
+    Api::V2::Models.communication
+  end
 end
