+++ config/routes.rb
@@ -41,11 +41,6 @@ Rails.application.routes.draw do
       post 'sync', to: 'appointments#sync_from_user'
     end
 
-    scope '/communications' do
-      get 'sync', to: 'communications#sync_to_user'
-      post 'sync', to: 'communications#sync_from_user'
-    end
-
     scope '/medical_histories' do
       get 'sync', to: 'medical_histories#sync_to_user'
       post 'sync', to: 'medical_histories#sync_from_user'
@@ -62,6 +57,11 @@ Rails.application.routes.draw do
         post 'register', to: 'users#register'
         post '/:id/request_otp', to: 'users#request_otp'
         post '/me/reset_password', to: 'users#reset_password'
+      end
+
+      scope '/communications' do
+        get 'sync', to: 'communications#sync_to_user'
+        post 'sync', to: 'communications#sync_from_user'
       end
 
       concerns :sync_routes
@@ -85,6 +85,11 @@ Rails.application.routes.draw do
         post 'register', to: 'users#register'
         post '/:id/request_otp', to: 'users#request_otp'
         post '/me/reset_password', to: 'users#reset_password'
+      end
+
+      scope '/communications' do
+        get 'sync', to: 'communications#sync_to_user'
+        post 'sync', to: 'communications#sync_from_user'
       end
 
       concerns :sync_routes
